Visualizing likelihoods and confidence ellipses
Please go through this lecture notebook about visualizing likelihoods of tweets and the confidence ellipses for annotated  corpuses. The positive and negative likelihoods of a tweet can be interpreted as a pair of numerical features  extracted from raw text. The visual inspection of the features that will feed a model allows to better understand the problem.